app-emulation/modules: fix -mpreferred-stack-boundary error
The smpcall.c unit fails to build leading to the failure
described in https://bugs.gentoo.org/show_bug.cgi?id=559798#c63

compilation fails when SSE related user flags are set in CFLAGS.
This is same condition as described in bug report
https://bugs.gentoo.org/show_bug.cgi?id=492964#c9

In the kernel SSE code is generally disabled so -mpreferred-stack-boundary=3
is forced on even on 64 bit systems. If any user CFLAGS requires
SSE instructions than the requirement for gcc becomes -mpreferred-stack-boundary=4
leading to a compilation failure.

Temporary fix until this is embedded in linux-mod.eclass because
it's not specific to vmware, I have opened the bug report #571800.