Unity support has been removed with version 12.
I have noticed warnings in the appLoader log:

  W115: Unable to map /opt/vmware/lib/vmware/lib/libvmware-unity-helper.so/libvmware-unity-helper.so

The library is indeed missing. It seems that the unity helper is not
anymore needed because upstream has removed unity support for Linux
guests and hosts (https://communities.vmware.com/thread/518735). The
release notes file confirms:
http://pubs.vmware.com/Release_Notes/en/workstation/12pro/workstation-12-release-notes.html

After removing the helper the following error is still present

  I125: Unable to find /opt/vmware/lib/vmware/bin/vmware-unity-helper in attempt to launch daemon.

but everything seems to work.