Remove from unbundling also libgksu2.so.0
libgksu2.so.0 seems to be a patched version of version 2.0.12. vmware
devs have added at least a symbol gksu_context_set_on_fork needed
to enter the serial number for vmware-workstation. I'm curing at the
same way also the vmware-player to keep the two ebuilds aligned to easy
the maintenance.

Thanks to Alexander Bezrukov for the report to be found in
https://bugs.gentoo.org/show_bug.cgi?id=559798#c86