Avoid preserve-libs with bundled libgksu2.so.0
With commit c6d45fdfc7ba3f7c7459c8e2369bb294681c9ffb it was necessary
to leave bundled libgksu2.so.0. Nevertheless the library libvmware-gksu.so
declares libgksu2.so.0 as DT_NEEDED even when it is using the bundled
version without setting RPATH correctly. Patch the lib to avoid portage
preserve-libs mechanism to be triggered when a system lib is available
(but not required).

Thanks to Zac Medico (https://bugs.gentoo.org/show_bug.cgi?id=215242#c32) for
the hint and to Luca Santarelli for the report
(https://github.com/gentoo/vmware/pull/15#issuecomment-197814568).