app-emulation/vmware-workstation: other workarounds for bundled-libs
Commit 052215b3c1e4b532b9d7a7872a0f46ecdb10f4cb has drastically reduced the
amount of not needed deps... introducing other problems.

1) The removal of gnome-base/librsvg exposes a problem with
libconf/lib/gtk-2.0/2.10.0/loaders/svg_loader.so (only with
bundled-libs), which is not able to find the bundled librsvg-2.so.2.

2) Some binaries (bin/vmware-vmx*) depends on libXinerama.so.1 and libXcursor.so.1
but with bundled-libs they are not able to find the embedded version and
fail if those libs are not installed in the system. patchelf --set-rpath doesn't work
with bin/vmware-vmw-debug so I moved the deps on libXinerama and libXcursor
to be always set.

Moreover, with gcc-5, bundled-libs is mandatory otherwise there is a problem
with symbols (https://bugs.gentoo.org/show_bug.cgi?id=578070#c8). Even
with VMWARE_USE_SHIPPED_LIBS set, appLoader tries to dynamically load at
runtime the best version of glib and fontconfig leading to a mix of
system/bundled libs. Let's fix some rpath to be sure that with
bundled-libs only the libs embedded with vmware are used!

Tested both on amd64 and ~amd64 systems.