removed USE flag bundled-libs
In portage a lot of deps are not anymore present because 12.5.9.7535481
version is more than two years old. The old USE=-bundled-libs configuration
is then not anymore working leading to an useless installation. Now the
default is an installation with bundled libs without options. There are of
course security issues but let's leave to the user the option to install the
software he paid for :-)

Signed-off-by: Fabio Rossi <rossi.f@inwind.it>